Data_Science_Projects/ML_Flask/requirements.txt
pandas==0.25.3
numpy==1.18
gensim==3.8.0
nltk==3.4.5
scikit-learn==0.22.1
joblib==0.14.1
pyarrow==0.17.0
gunicorn==20.0.4
flask==1.1.1
snowflake-connector-python==2.2.9